+++ pybop/__init__.py
@@ -64,6 +64,7 @@ from .transformation.transformations import (
     ScaledTransformation,
     LogTransformation,
     ComposedTransformation,
+    UnitHyperCube,
 )
 
 #
@@ -129,6 +130,7 @@ from .experimental._evaluation import SequentialJaxEvaluator
 
 from .optimisers._cuckoo import CuckooSearchImpl
 from .optimisers._adamw import AdamWImpl
+from .optimisers._gradient_descent import GradientDescentImpl
 from .optimisers.base_optimiser import BaseOptimiser, OptimisationResult
 from .optimisers.base_pints_optimiser import BasePintsOptimiser
 from .optimisers.scipy_optimisers import (